prakriti_backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
aiosqlite
greenlet
pydantic
pydantic-settings
redis[hiredis]
firebase-admin
python-jose[cryptography]
passlib[bcrypt]
python-multipart
fpdf2
httpx
alembic
python-dotenv
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
aiosqlite
greenlet
pydantic
pydantic-settings
redis[hiredis]
firebase-admin
anthropic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
fpdf2
httpx
alembic
path-lib
python-dotenv